+++ src/Symfony/Component/Security/CHANGELOG.md
@@ -1,6 +1,12 @@
 CHANGELOG
 =========
 
+3.4.0
+-----
+
+ * Using voters in the `AccessDecisionManager` without `VoterInterface` is now
+   deprecated and this functionality will be removed in 4.0.
+
 3.3.0
 -----